Fixed validation of at least 2 items for radio/checkobxo when JS enabled
