+++ designer/client/src/javascripts/types.js
@@ -1,12 +1,4 @@
 /**
- * @typedef {object} ListItem
- * @property {string} [text] - text
- * @property {{ text?: string }} [hint] - hint
- * @property {string} [value] - value
- * @property {string} [id] - id
- */
-
-/**
  * @typedef {object} FormDetails
  * @property {Element} saveButton - button
  * @property {Element} cancelButton - button
@@ -20,4 +12,8 @@
  * @property {ListItem[]} listItems - array of list items
  * @property {HTMLElement} listItem - single list item
  * @property {string} baseClassName - base class name
+ */
+
+/**
+ * @import { ListItem } from '@defra/forms-model'
  */
