+++ designer/server/src/routes/forms/editor-v2/question-details-helper.js
@@ -19,7 +19,7 @@ const listUniquenessSchema = Joi.object({
 })
 
 /**
- * @param { { id?: string, text?: string, hint?: { text: string }, value?: string } | undefined } itemForEdit
+ * @param { ListItem | undefined } itemForEdit
  * @param {boolean} expanded
  */
 export function setEditRowState(itemForEdit, expanded) {
@@ -184,7 +184,7 @@ export function handleEnhancedActionOnPost(request, stateId, questionDetails) {
 }
 
 /**
- * @import { ComponentDef,  FormEditorInputQuestionDetails, QuestionSessionState, FormDefinition, ListComponentsDef } from '@defra/forms-model'
+ * @import { ComponentDef,  FormEditorInputQuestionDetails, QuestionSessionState, ListItem } from '@defra/forms-model'
  * @import { Request, RequestQuery } from '@hapi/hapi'
  * @import { Yar } from '@hapi/yar'
  */
