+++ model/src/form/form-editor/types.ts
@@ -306,6 +306,15 @@ export type FormEditorInputQuestionDetails = Pick<
   | 'jsEnabled'
 >
 
+export interface ListItem {
+  text?: string
+  hint?: {
+    text?: string
+  }
+  value?: string
+  id?: string
+}
+
 export interface QuestionSessionState {
   questionType?: ComponentType
   questionDetails?: Partial<ComponentDef>
@@ -316,12 +325,7 @@ export interface QuestionSessionState {
     radioValue?: string
     expanded?: boolean
   }
-  listItems?: {
-    text?: string
-    hint?: { id?: string; text: string }
-    value?: string
-    id?: string
-  }[]
+  listItems?: ListItem[]
 }
 
 export interface GovukField {
